Test of fit statbility
Implemented a simple test of fit stability and produced a plot. I sample a subset of the points in a continium window and perform a fit, and repeat this 5000 times to get an estimate of the fit variance. I also do this for a range of sizes of the subset.

Also renamed the automatic methods for selecting background windows as well as adding a few convenience functions.