Added functions to compare contniua fit on data
Using the different window definitions gives a set of continua fits for all lines in the frame,
to compare the performance of the different definitions their central continuum value are standardised
around zero. This is done using either the ensamble average, or the value of the pixel in the column
containing the highest pixel in the convolved FTS spectra, or the average that pixel with its five
neigthbours at either side.